+++ CHANGELOG.md
@@ -7,6 +7,12 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+## [0.1.0] - 2023-02-09
+
+Extra argument enables more "group labels" than just job and instance.
+
+This release is backwards-compatible with portertech 0.0.2
+
 ## [0.0.2] - 2020-08-19
 
 ### Changed